// Source Code/AdminOnlineFloralDelivery/src/app/models/bouquetevent.model.ts
export class BouquetEvent{
  bouquetId?: number;
  bouquetName?: string;
  description?: string;
  price?: number;
  secondPrice?:number;
  eventPrice?: number;
  quantity?:number;
  created?:string;
  categoryId?: number;
  categoryName?: string;
  status: boolean;
  imageUrl: string;
  eventId: number;
  eventName: string;
}